requirements.txt
aiohttp
chardet
discord.py==1.2.3
feedparser
websockets
